comments

+++ src/main/java/com/example/otrtesttask/service/BranchService.java
@@ -2,7 +2,9 @@ package com.example.otrtesttask.service;
 
 import com.example.otrtesttask.exceptions.CustomApiException;
 import com.example.otrtesttask.jooq.tables.pojos.Branch;
+import com.example.otrtesttask.jooq.tables.pojos.Employee;
 import com.example.otrtesttask.repository.BranchRepository;
+import com.example.otrtesttask.repository.EmployeeRepository;
 import org.jooq.Condition;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,12 +16,18 @@ import java.util.List;
 public class BranchService {
     @Autowired
     private BranchRepository branchRepository;
+    @Autowired
+    private EmployeeRepository employeeRepository;
 
     public Branch create(Branch branch) throws CustomApiException {
+
+        // Если передали id, то возвращаем ошибку
         if (branch.getId() != null)
             throw new CustomApiException("Id field is prohibited", HttpStatus.BAD_REQUEST);
+        // В запросе не передан title
         if (branch.getTitle() == null)
             throw new CustomApiException("Missing required field: title", HttpStatus.BAD_REQUEST);
+
         return branchRepository.insert(branch);
     }
 
@@ -42,6 +50,11 @@ public class BranchService {
     }
 
     public Boolean delete(Integer id) throws CustomApiException {
+
+        List<Employee> employees = employeeRepository.findEmployeesByBranchId(id);
+        if (employees != null)
+            throw new CustomApiException(String.format("Some employees are attached to branch with %d", id), HttpStatus.BAD_REQUEST);
+
         Boolean b = branchRepository.delete(id);
         if (!b)
             throw new CustomApiException(String.format("Branch with id %d not found", id), HttpStatus.NOT_FOUND);
