+++ src/main/java/com/example/otrtesttask/service/PositionService.java
@@ -1,7 +1,9 @@
 package com.example.otrtesttask.service;
 
 import com.example.otrtesttask.exceptions.CustomApiException;
+import com.example.otrtesttask.jooq.tables.pojos.Employee;
 import com.example.otrtesttask.jooq.tables.pojos.Position;
+import com.example.otrtesttask.repository.EmployeeRepository;
 import com.example.otrtesttask.repository.PositionRepository;
 import org.jooq.Condition;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,11 +17,17 @@ public class PositionService {
     @Autowired
     private PositionRepository positionRepository;
 
+    @Autowired
+    private EmployeeRepository employeeRepository;
+
     public Position create(Position position) throws CustomApiException {
+        // Если передали id, то возвращаем ошибку
         if (position.getId() != null)
             throw new CustomApiException("Id field is prohibited", HttpStatus.BAD_REQUEST);
+        // В запросе не передан title
         if (position.getTitle() == null)
             throw new CustomApiException("Missing required field: title", HttpStatus.BAD_REQUEST);
+
         return positionRepository.insert(position);
     }
 
@@ -35,6 +43,11 @@ public class PositionService {
     }
 
     public Boolean delete(Integer id) throws CustomApiException {
+
+        List<Employee> employees = employeeRepository.findEmployeesByBranchId(id);
+        if (employees != null)
+            throw new CustomApiException(String.format("Some employees are attached to position with %d", id), HttpStatus.BAD_REQUEST);
+
         Boolean b = positionRepository.delete(id);
         if (!b)
             throw new CustomApiException(String.format("Branch with id %d not found", id), HttpStatus.NOT_FOUND);
